Get user credentials from Github
As part of the authentication flow, we'll use the Github user ID number
to determine if the user is new and needs to be created, or if they just
are logging in again.  We'll also use their Github profile to
pre-populate some of their information in the database, like an avatar
URL and email address.